pyomyo/requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.2
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pygame==2.6.1
pynput==1.7.7
pyobjc-core==10.3.2
pyobjc-framework-ApplicationServices==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-CoreText==10.3.2
pyobjc-framework-Quartz==10.3.2
pyomyo==0.0.5
pyparsing==3.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
scikit-learn==1.5.2
scipy==1.14.1
six==1.17.0
threadpoolctl==3.5.0
xgboost==2.1.3
